CDPT-1511 Move moj-auth to own folder to improve readability.
