+++ composer.json
@@ -76,7 +76,8 @@
         "ext-posix": "*",
         "ext-mysqli": "*",
         "ext-zlib": "*",
-        "league/oauth2-client": "^2.7"
+        "league/oauth2-client": "^2.7",
+        "firebase/php-jwt": "^6.10"
     },
     "require-dev": {
         "squizlabs/php_codesniffer": "^3.0.2"
